requirements/base.txt
Django==6.0.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==25.1
Pillow==11.2.1
psycopg2-binary==2.9.10
django-storages==1.14.6
boto3==1.37.37
python-decouple==3.8
